% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCommonLR.R
\name{calcCommonLR}
\alias{calcCommonLR}
\title{calcCommonLR}
\usage{
calcCommonLR(mlefit1, mlefit2, DC1 = NULL, DC2 = NULL, returnPerMarker = FALSE)
}
\arguments{
\item{mlefit1}{1st Fitted object using calcMLE/contLikMLE function.}

\item{mlefit2}{2nd Fitted object using calcMLE/contLikMLE function.}

\item{DC1}{Optional deconvolute object from mlefit1}

\item{DC2}{Optional deconvolute object from mlefit2}

\item{returnPerMarker}{Whether returning LR per marker instead of across all.}
}
\value{
A matrix of LR between all unknown pairs. NULL if some have no unknowns. A ists if returnPerMarker=TRUE
}
\description{
Calculating the LR for common unknowns
}
\details{
LR calculation of whether two paired unknowns between two fitted models is identical (or not).
Based on formula by Slooten (2017): Identifying common donors in DNA mixtures, with applications to database searches
}
\author{
Oyvind Bleka
}
