% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGjoint.R
\name{calcGjoint}
\alias{calcGjoint}
\title{calcGjoint}
\usage{
calcGjoint(freq, nU = 1, fst = 0, refK = NULL, refR = NULL,
  ibd = c(1, 0, 0), sortComb = FALSE)
}
\arguments{
\item{freq}{A vector of allele frequencies for a given population.}

\item{nU}{Number of unknowns (where the first is a related)}

\item{fst}{Assumed theta/fst-correction}

\item{refK}{contains a vector of alleles for the known typed reference profiles (a,b,c,d...)}

\item{refR}{contains a vector of alleles for a related reference profile (a,b)}

\item{ibd}{the identical by decent coefficients of the relationship denotation}

\item{sortComb}{boolean of whether only the outcome G1,G2>=G3>=...>=GnU should be calculated (the Gprobs are symmetrical for k=2,..,U)}
}
\value{
Glist A list with genotypes and genotype probabilities
}
\description{
getGlist Returns a list of joint genotypes with corresponding joing probabilities for unknowns contributors for a given marker.
}
\details{
The function returns the list of all possible genotypes, with corresponding probabilities. The allele-names in popFreq needs to be numbers.
}
\examples{
freq = rgamma(8,1,1)
freq = freq/sum(freq) 
names(freq) <- 1:length(freq)
system.time({ foo = calcGjoint(freq,nU=3,fst=0.1,refK=c("2","3","1","1"),refR=c("2","3"),ibd=c(1/4,1/2,1/4),sortComb=FALSE) })
system.time({ foo = calcGjoint(freq,nU=3,fst=0.1,refK=c("2","3","1","1"),refR=c("2","3"),ibd=c(1/4,1/2,1/4),sortComb=TRUE) })
}
\author{
Oyvind Bleka
}
