% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRint.R
\name{calcLRint}
\alias{calcLRint}
\title{calcLRint}
\usage{
calcLRint(
  mlefitHp,
  mlefitHd,
  reltol = 0.001,
  maxEval = 0,
  dev = 3,
  verbose = TRUE
)
}
\arguments{
\item{mlefitHp}{A fitted object returned from contLikMLE (under Hp)}

\item{mlefitHd}{A fitted object returned from contLikMLE (under Hd)}

\item{reltol}{Required relative tolerance error of evaluations in integration routine. Default is 0.001.}

\item{maxEval}{Maximum number of evaluations in the adaptIntegrate function. Default is 0 which gives an infinite limit.}

\item{dev}{Scaling relative to 2*SE (i.e. 97.5 percentile). Should cover most of post}

\item{verbose}{Whether printing simulation progress. Default is TRUE}
}
\value{
returned inferred LR with error range (reflects relative error)
}
\description{
Calculating marginal likelihood based on integration
}
\details{
The procedure is a wrapper for the numerical integration function.
}
\author{
Oyvind Bleka
}
