% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRmcmc.R
\name{calcLRmcmc}
\alias{calcLRmcmc}
\title{calcLRmcmc}
\usage{
calcLRmcmc(
  mlefitHp,
  mlefitHd,
  niter = 2000,
  delta = 2,
  quantile = 0.05,
  seed = NULL,
  accRate = 0.25,
  accRateTol = 0.1,
  niterTune = 200,
  diffSeed = 999,
  verbose = TRUE,
  traceplot = TRUE,
  mcmcObjList = NULL
)
}
\arguments{
\item{mlefitHp}{A fitted object returned from contLikMLE (under Hp)}

\item{mlefitHd}{A fitted object returned from contLikMLE (under Hd)}

\item{niter}{Number of samples in the MCMC-sampling.}

\item{delta}{A numerical parameter to scale with the covariance function Sigma. Default is 2. Will be tuned automtically.}

\item{quantile}{The quantile used to report conservative LR}

\item{seed}{The user can set seed if wanted}

\item{accRate}{aimed acceptance rate}

\item{accRateTol}{Difference tolerance regarding acceptance rate}

\item{niterTune}{Number of samples for tuning delta in the MCMC-sampling.}

\item{diffSeed}{Seed difference}

\item{verbose}{Whether progress should be printed}

\item{traceplot}{Whether to add a traceplot with 95\% CI}

\item{mcmcObjList}{A list with hp/hd earlier run with contLikMCMC output}
}
\value{
A list with different LR statistics from the MCMC based simulations
}
\description{
A function for calculating LR from MCMC simulations
}
\details{
Returns a conservative (quantile) LR or a Full Bayesian based LR
}
\author{
Oyvind Bleka
}
