% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRmle.R
\name{calcLRmle}
\alias{calcLRmle}
\title{calcLRmle}
\usage{
calcLRmle(mlefitHp, mlefitHd)
}
\arguments{
\item{mlefitHp}{A fitted object returned from calcMLE (under Hp)}

\item{mlefitHd}{A fitted object returned from calcMLE (under Hd)}
}
\description{
A function for summarizing LR (MLE)
}
