% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRupper.R
\name{calcLRupper}
\alias{calcLRupper}
\title{calcLRupper}
\usage{
calcLRupper(POIidx, mle, scale = TRUE)
}
\arguments{
\item{POIidx}{Index of POI (in refData)}

\item{mle}{Fitted object using calcMLE (under Hd)}

\item{scale}{Whether to scale the RMP when having more than 2 unknowns (recommended)}
}
\value{
Returning theoretical upper boundary LR (log10 scale)
}
\description{
Obtaining upper boundary LR for POI
}
\details{
The function takes a fitted MLE object under Hd (where hypothesis$knownRef is the POI)
}
