% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQual.R
\name{calcQual}
\alias{calcQual}
\title{calcQual}
\usage{
calcQual(evids, knownContr, knownNonContr, nUnknowns, theta, prD, prC, freq)
}
\arguments{
\item{evids}{vector of alleles present at a given locus for any number of replicates. Same as 'Repliste' in likEvid function.}

\item{knownContr}{vector of genotypes for the known contributors. Genotype 12/17 should be given as a vector c(12,17) and genotypes 12/17,14/16, should be given as a unique vector: c(12,17,14,16). If empty, set to 0.}

\item{knownNonContr}{vector of genotypes for the known non-contributors}

\item{nUnknowns}{Number of unknown individuals under H. Set to 0 if there are no unknown contributors.}

\item{theta}{theta correction, value must be taken in [0,1)}

\item{prD}{probability of heterozygotes dropout. It is possible to assign different values per contributor.}

\item{prC}{probability of drop-in applied per locus}

\item{freq}{vector of the corresponding allele frequencies of the analysed locus in the target population}
}
\value{
likelihood value
}
\description{
Implementation of the qualitative likelihood calculations
}
\details{
Assumes that prD is a identical for all contributors. Otherwise need to run forensim::likEvid (or possibly other implementations)
From likEvid documentation: If there are two replicates, showing alleles 12,13, and 14 respectively, then evids should be given as c(12,13,0,14), where the 0 is used as a separator. An empty replicate is simply 0. For example, replicates (12,13) and and one empty replicate must be given as: c(12,14,0,0).
}
