% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRMPfst.R
\name{calcRMPfst}
\alias{calcRMPfst}
\title{calcRMPfst}
\usage{
calcRMPfst(dat, POIind = 1, condInd = NULL, fst = 0)
}
\arguments{
\item{dat}{A object retured from getData function}

\item{POIind}{Index of the references being Person of interest (POI)}

\item{condInd}{Index of conditional references}

\item{fst}{The co-ancestry coefficient (theta-correction). Can be a vector (must contain the marker names)}
}
\value{
ret A vector with random match probabilities for each markers
}
\description{
A helpfunction for calculating the random match probability for a person of interest (given fst)
}
\author{
Oyvind Bleka
}
