% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTippet.R
\name{calcTippet}
\alias{calcTippet}
\title{calcTippet}
\usage{
calcTippet(
  tipIdx,
  mlefitHp,
  mlefitHd,
  niter = 100,
  type = "MLE",
  LRobs = NULL,
  MLEopt = NULL,
  INTopt = NULL,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{tipIdx}{Index in condOrder which are replaced with a random man}

\item{mlefitHp}{A fitted object returned from contLikMLE (under Hp)}

\item{mlefitHd}{A fitted object returned from contLikMLE (under Hd)}

\item{niter}{Number of non-contributor iterations}

\item{type}{type of Tippet analysis to conduct (MLE or INT)}

\item{LRobs}{The user can send an observed LR to superimpose to the plot}

\item{MLEopt}{MLE options (must include nDone, steptol, delta, difftol, seed is optional)}

\item{INTopt}{INT options (must include reltol, maxEval, dev)}

\item{seed}{Seed used for reproducibility. Note that seed can also be a variable in MLEopt option}

\item{verbose}{Whether printing simulation progress. Default is TRUE}
}
\value{
returned non-contributor LR values
}
\description{
Performing tippet analysis for a given model
}
\details{
The procedure is a wrapper for the numerical integration function.
}
\author{
Oyvind Bleka
}
