% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contLikSearch.R
\name{contLikSearch}
\alias{contLikSearch}
\title{contLikSearch}
\usage{
contLikSearch(NOC = 2:3, modelDegrad = FALSE, modelBWstutt = FALSE,
  modelFWstutt = FALSE, samples = NULL, popFreq = NULL,
  refData = NULL, condOrder = NULL, knownRefPOI = NULL, prC = 0,
  nDone = 2, threshT = 50, fst = 0, lambda = 0, pXi = function(x)
  1, delta = 1, kit = NULL, verbose = TRUE, maxIter = 100,
  knownRel = NULL, ibd = c(1, 0, 0), pXiFW = function(x) 1,
  seed = NULL, maxThreads = 32, alpha = 0.01)
}
\arguments{
\item{NOC}{#The vector of number of contributors to search}

\item{modelDegrad}{Boolean of whether degradation should be modeled (requires kit to be specified): can be vector (FALSE,TRUE)}

\item{modelBWstutt}{Boolean of whether backward stutter should be modeled: can be vector (FALSE,TRUE)}

\item{modelFWstutt}{Boolean of whether forward stutter should be modeled: can be vector (FALSE,TRUE)}

\item{samples}{A List with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with (2-size) allele-vector given in list element [[i]][[s]].}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model.}

\item{knownRefPOI}{Specify known POI under Hp, under non-contributing references (under Hd) from refData (index). For instance knownRef=(1,2) means that reference 1 and 2 is known non-contributor in the hypothesis. This affectes coancestry correction.}

\item{prC}{A numeric for allele drop-in probability. Default is 0.}

\item{nDone}{Number of optimizations required providing equivalent results (same logLik value obtained)}

\item{threshT}{The detection threshold given. Used when considering probability of allele drop-outs.}

\item{fst}{The co-ancestry coefficient. Default is 0.}

\item{lambda}{Parameter in modeled peak height shifted exponential model. Default is 0.}

\item{pXi}{Prior function for xi-parameter (stutter). Flat prior on [0,1] is default.}

\item{delta}{Scaling of variation of normal distribution when drawing random startpoints. Default is 1.}

\item{kit}{Used to model degradation. Must be one of the shortnames of kit: check getKit()}

\item{verbose}{Boolean whether printing optimization progress. Default is TRUE.}

\item{maxIter}{Maximum number of iterations for the optimization. Default is 100.}

\item{knownRel}{gives the index of the reference which the 1st unknown is related to.}

\item{ibd}{the identical by decent coefficients of the relationship (specifies the type of relationship)}

\item{pXiFW}{Prior function for xiFW-parameter (FW stutter). Flat prior on [0,1] is default.}

\item{seed}{The user can set seed if wanted}

\item{maxThreads}{Maximum number of threads to be executed by the parallelization}

\item{alpha}{The significance level used for the envelope test (validMLEmodel). Default is 0.01}
}
\value{
ret A list(fit,model,nDone,delta,seed,prepareC) where fit is Maximixed likelihood elements for given model.
}
\description{
contLikSearch search through the models  optimizes the likelihood function of the DNA mixture model
}
\details{
Function calls the likelihood function implemented in C++ which uses the package Boost and paralellisation using OpenMP
}
\author{
Oyvind Bleka
}
