% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{deconvolve}
\usage{
deconvolve(
  mlefit,
  alpha = 0.95,
  maxlist = 1000,
  signif = 4,
  checkCalcs = FALSE
)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE function.}

\item{alpha}{Required sum of the listed posterior probabilities.}

\item{maxlist}{The ranked deconvolved profile list will not exceed this number (used to avoid endless search).}

\item{signif}{Number of significant numbers}

\item{checkCalcs}{Return only calculated log-likelihood per markers (summed). Used for testing.}
}
\value{
ret A list(table1,table2,table3,table4,rankGi,rankG,pG) where rankG is the ranked genotypes with corresponding probabilities in pG. rankgGi is the same, but per marker. table1 is rankG and pG combined (joint results). table2 uses rankGi to find marginal results for top-genotypes. table3 and table4 shows this marginalized on genotypes and alleles per contributor
}
\description{
deconvolve ranks the set of the most conditional posterior probability of genotypes the STR DNA mixture given a fitted model under a hypothesis.
}
\details{
The procedure calculates the likelihood for each single locus. Then it combines the most probable genotypes from each loci to produce a ranked list of deconvolved profiles.
}
\author{
Oyvind Bleka
}
