% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm.R
\name{efm}
\alias{efm}
\title{efm}
\usage{
efm(envirfile = NULL)
}
\arguments{
\item{envirfile}{A Rdata file including a saved environment of a project  (variable must be named mmTK)}
}
\description{
efm is a GUI wrapper for EuroForMix
}
\details{
The function is a graphical layer for the functions in the package euroformix. See ?euroformix for more information.
}
\author{
Oyvind Bleka
}
