% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm_createReport.R
\name{efm_createReport}
\alias{efm_createReport}
\title{efm_createReport}
\usage{
efm_createReport(mmTK, type = "EVID", save = TRUE)
}
\arguments{
\item{mmTK}{An environment created with the efm function (can be loaded from project file)}

\item{type}{Indicate type of report to store (EVID or DC/DB)}

\item{save}{whether to save to text-file}
}
\description{
Creating report from stored efm session
}
\details{
The function is a graphical layer for the functions in the package euroformix. See ?euroformix for more information.
}
\author{
Oyvind Bleka
}
