% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm_DBsearch.R
\name{efm_DBsearch}
\alias{efm_DBsearch}
\title{efm_DBsearch}
\usage{
efm_DBsearch(mmTK, ITYPE = "QUAN")
}
\arguments{
\item{mmTK}{The EFM GUI environment variable (CONTAINS DATA AND MODEL FOR CALCULATIONS)}

\item{ITYPE}{Type of LR interpretation (QUAN=EFM or QUAL=LRmix)}
}
\description{
A EFM GUI helpfunction to run the Database search LR comparisons
}
\author{
Oyvind Bleka
}
