% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm_gfile.R
\name{efm_gfile}
\alias{efm_gfile}
\title{efm_gfile}
\usage{
efm_gfile(text, type, filter = list(), initf = NULL)
}
\arguments{
\item{text}{Input argument to gfile}

\item{type}{Input argument to gfile}

\item{filter}{Input argument to gfile}

\item{initf}{Input argument to gfile}
}
\description{
A helpfunction to obtain file names from browser
}
\details{
Ensures that correct coding is applied to the file name returned from gWidgets2::gfile
}
\author{
Oyvind Bleka
}
