% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDataFromProject.R
\name{exportDataFromProject}
\alias{exportDataFromProject}
\title{exportDataFromProject}
\usage{
exportDataFromProject(
  projectFile,
  evidfn = "evids.csv",
  reffn = "refs.csv",
  freqfn = "freqs.csv"
)
}
\arguments{
\item{projectFile}{File name (full path) of the project file}

\item{evidfn}{File name for the exported evidence data}

\item{reffn}{File name for the exported reference data}

\item{freqfn}{File name for the exported frequency data}
}
\description{
Exports data from a EFM-project file into multiple files in EFM format
}
\author{
Oyvind Bleka
}
