% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgammamodel.R
\name{fitgammamodel}
\alias{fitgammamodel}
\title{fitgammamodel}
\usage{
fitgammamodel(y, x = NULL, niter = 10, delta = 1, plott = FALSE,
  alpha = 0.05, offset = 125, scale = 100, restrictDeg = TRUE)
}
\arguments{
\item{y}{Vector with sum peak heights (per loci)}

\item{x}{Vector with fragment lengths (base pair) (only provided for degradation)}

\item{niter}{Number of random samples}

\item{delta}{Standard deviation of normal distribution when drawing random startpoints. Default is 1.}

\item{plott}{Whether plotting fitted model along with data}

\item{alpha}{The alpha/2 and 1-alpha/2 percentiles will be plotted if x is provided}

\item{offset}{The shift of bp in degradation model.}

\item{scale}{The scaling of shifted bp in degradation model}

\item{restrictDeg}{Whether to restrict degradation param to be less than 1}
}
\value{
Estimated parameters (mean,coefVar,slopePara)
}
\description{
A function to fit sumPH of the observed peak heights to a gamma model
}
\details{
This function is used for fitting a gamma regression model
}
\examples{
\dontrun{
th = c(1000,0.8,0.6)
n = 100 #number of samples
x = seq(10,300,l=n)
y = rgamma(n,shape=2/th[2]^2*th[3]^((x-125)/100),scale=th[1]*th[2]^2)
fitgammamodel(y,x,plott=TRUE,alpha=0.05)
fitgammamodel(y,x,plott=TRUE,alpha=0.05)
}
}
\author{
Oyvind Bleka
}
