% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqImport.R
\name{freqImport}
\alias{freqImport}
\title{freqImport}
\usage{
freqImport(f = NULL, url = FALSE, xml = FALSE)
}
\arguments{
\item{f}{The path of file(s) or an url. If file(s), then the url argument must be FALSE.}

\item{url}{Whether f is pointing to an url or not. Default is FALSE.}

\item{xml}{Whether f is pointing to an xml format file or not. If FALSE then table formats with standard separator types are considered. Default is FALSE.}
}
\value{
dblist A list giving the population frequencies for every population
}
\description{
Imports allele frequencies directly from file(s) or URLs
}
\details{
The function supports XML as input, directly from an URL or local file (for instance from the STRidER webpage).
}
\author{
Oyvind Bleka
}
