% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs_listToTable.R
\name{freqs_listToTable}
\alias{freqs_listToTable}
\title{freqs_listToTable}
\usage{
freqs_listToTable(popFreq)
}
\arguments{
\item{popFreq}{List with frequencies per marker}
}
\value{
Table with alleles ordered. Ready to be exported to frequency file.
}
\description{
Converting frequency data from list to table format
}
\author{
Oyvind Bleka
}
