% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDataset.R
\name{genDataset}
\alias{genDataset}
\title{genDataset}
\usage{
genDataset(
  nC,
  popFreq,
  mu = 1000,
  sigma = 0.1,
  sorted = FALSE,
  threshT = 50,
  refData = NULL,
  mx = NULL,
  nrep = 1,
  stutt = 0,
  prC = 0,
  lambda = 0,
  beta = 1,
  kit = NULL,
  stuttFW = 0
)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{popFreq}{A list with allele frequencies for a given population.}

\item{mu}{Expected peak heights for a het. single contributor allele}

\item{sigma}{Coeffecient of variance of peak heights.}

\item{sorted}{Wheter sorting the contributors with respect to decreasingly mixture proportions.}

\item{threshT}{Required allele peak height in mixture (can be a vector with names giving the loci names)}

\item{refData}{A list with given reference profiles given as refData[[i]][[s]]. Default is random from population.}

\item{mx}{A vector with known mixture proportions. Default is random uniform.}

\item{nrep}{Number of peak height replicates (same contributors) to generate. Default is 1.}

\item{stutt}{A numerical stutter proportion (n-1). Default is 0.}

\item{prC}{A numerical dropin probability (can be a vector with names giving the loci names). Default is 0.}

\item{lambda}{The rate parameter in the exponential distribution for simulating drop-in peak heights (can be a vector with names giving the loci names). Default is 0.}

\item{beta}{The degradation slope parameter used for simulating degradation trend (requires valid kit to be specified). Default is 1.}

\item{kit}{shortname of kit: {"ESX17","ESI17","ESI17Fast","ESX17Fast","Y23","Identifiler","NGM","ESSPlex","ESSplexSE","NGMSElect","SGMPlus","ESX16", "Fusion","GlobalFiler"}}

\item{stuttFW}{A numerical Forward stutter proportion (n+1). Default is 0.}
}
\value{
List with elements theta,samples,refData where theta is the true parameters of the model. samples is a list with samples which for each samples has locus-list elements with list elements adata and hdata
}
\description{
Function for generating mixture samples for given model parameters
}
\details{
genDataset samples random mixture peak heights given as gamma(rho*sum(h_k),tau), with h_k as peak height of k-te contributor.
genData conditions on alleles given by refData. Empty references are generated with population frequencies.
}
\examples{
\dontrun{ 
kit = "SGMPlus"
popfile = paste(path.package("euroformix"),"FreqDatabases",
 paste0(kit,"_Norway.csv"),sep=.Platform$file.sep)
popFreq = freqImport(popfile)[[1]] #obtain list with population frequencies
gen = genDataset(2,popFreq,beta=.7,kit=kit,stutt=.1,stuttFW=.05)
plotEPG2(gen$samples,kit = kit,refData=gen$refData,AT=50) #visualize samples
}
}
\author{
Oyvind Bleka
}
