% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGlist.R
\name{getGlist}
\alias{getGlist}
\title{getGlist}
\usage{
getGlist(popFreq, fst = 0, refKlist = NULL, refRlist = NULL,
  ibd = c(1, 0, 0))
}
\arguments{
\item{popFreq}{A list of allele frequencies for a given population. Each named element in the list must be a allele-named vector with allele frequencies.}

\item{fst}{Assumed theta/fst-correction}

\item{refKlist}{contains a list of vector of alleles for the known typed reference profiles (a,b,c,d...)}

\item{refRlist}{contains a list of vector of alleles for a related reference profile (a,b)}

\item{ibd}{the identical by decent coefficients of the relationship denotation}
}
\value{
Glist A list with genotypes and genotype probabilities for each locus.
}
\description{
getGlist Returns a list of genotypes with corresponding genotype probabilities for an unknown individual.
}
\details{
The function returns the list of all possible observed genotypes for each marker. A wrapper of the function calcGjoint.
}
\author{
Oyvind Bleka
}
