% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarkerVal.R
\name{getMarkerVal}
\alias{getMarkerVal}
\title{getMarkerVal}
\usage{
getMarkerVal(vec, marker = "")
}
\arguments{
\item{vec}{A vector with values to sort by letting names(vec) have same order as orderedNames}

\item{marker}{The marker to extract value of}
}
\value{
the value for selected marker
}
\description{
Extract value with respect to names of an input (case insensitive)
}
\details{
Helpfunction to extract a specific value from a vector wrt a specific locus 
Used in functions genDataset, prepareData
}
\author{
Oyvind Bleka
}
