% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamLimits.R
\name{getParamLimits}
\alias{getParamLimits}
\title{getParamLimits}
\usage{
getParamLimits(mlefit, dev = 2)
}
\arguments{
\item{mlefit}{Fitted object using calcMLE}

\item{dev}{Scaling relative to 2*SE (i.e. 97.5percentil). Should cover most of posterior distr outcome}
}
\value{
returned object from calcINT
}
\description{
Provides model parameter limitations
}
\details{
Used further for numerical integration
}
\author{
Oyvind Bleka
}
