% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleType.R
\name{getSampleType}
\alias{getSampleType}
\title{getSampleType}
\usage{
getSampleType(samples, kit = NULL, LUSsymbol = "_")
}
\arguments{
\item{samples}{A list of samples (evidence) with structure [[samplename]][[locus]] = list(adata,...)}

\item{kit}{Kitname returned from getKit()}

\item{LUSsymbol}{Symbol to be recognized as a LUS allele}
}
\description{
A function for determining data format
}
\details{
returns "EPG" (CE),"LUS" (MPS with LUS format) or "MPS" (else it is string)
}
