% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPG.R
\name{plotEPG}
\alias{plotEPG}
\title{plotEPG}
\usage{
plotEPG(Data, kitname, threshT = 0, refcond = NULL, showPH = FALSE)
}
\arguments{
\item{Data}{List of adata- and hdata-elements ("samples")}

\item{kitname}{Short name of kit: See supported kits with getKit()}

\item{threshT}{The detection threshold can be shown in gray in the plot.}

\item{refcond}{condition on a list$refname$locname$adata of reference alleles which are labeled in EPG}

\item{showPH}{Whether peak heights should be included in the plot}
}
\description{
EPG plotter created by Oskar Hansson.
}
\details{
Plots peak height with corresponding allele for one sample for a given kit.
}
\author{
Oyvind Bleka
}
