% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPG2.R
\name{plotEPG2}
\alias{plotEPG2}
\title{plotEPG2}
\usage{
plotEPG2(
  mixData,
  kit,
  refData = NULL,
  AT = NULL,
  ST = NULL,
  dyeYmax = TRUE,
  plotRepsOnly = TRUE,
  options = NULL
)
}
\arguments{
\item{mixData}{List object of [[sampleName]][[locus]] =list(adata,hdata), Allele vector adata (can be strings or numeric), intensity vector hdata (must be numeric)}

\item{kit}{Short name of kit: See supported kits with getKit()}

\item{refData}{List of refData[[reference]][[locus]] or refData[[locus]][[reference]] to label references (flexible). Visualizer will show dropout alleles.}

\item{AT}{A detection threshold can be shown in a dashed line in the plot (constant). Possibly a vector with locus column names}

\item{ST}{A stochastic threshold can be shown in a dashed line in the plot (constant). Possibly a vector with locus column names}

\item{dyeYmax}{Whether Y-axis should be same for all markers (FALSE) or not (TRUE this is default)}

\item{plotRepsOnly}{Whether only replicate-plot is shown in case of multiple samples (TRUE is default)}

\item{options}{A list of possible plot configurations. See comments below}
}
\value{
sub A plotly widget
}
\description{
EPG data visualizer (interactive)
}
\details{
Plots peak height with corresponding allele for sample(s) for a given kit.
}
\author{
Oyvind Bleka
}
