% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPS2.R
\name{plotMPS2}
\alias{plotMPS2}
\title{plotMPS2}
\usage{
plotMPS2(mixData, refData = NULL, AT = NULL, ST = NULL,
  grpsymbol = "_", locYmax = TRUE, options = NULL)
}
\arguments{
\item{mixData}{List of mixData[[ss]][[loc]] =list(adata,hdata), with samplenames ss, loci names loc, allele vector adata (can be strings or numeric), intensity vector hdata (must be numeric)}

\item{refData}{List of refData[[rr]][[loc]] or refData[[loc]][[rr]] to label references (flexible). Visualizer will show dropout alleles.}

\item{AT}{A detection threshold can be shown in a dashed line in the plot (constant). Possibly a AT[[loc]] list.}

\item{ST}{A stochastic threshold can be shown in a dashed line in the plot (constant). Possibly a ST[[loc]] list.}

\item{grpsymbol}{A separator for each allele giving plot grouping. Useful for separating conventional repeat units (RU) and sequence variant.}

\item{locYmax}{A boolean of whether Y-axis should be same for all markers (FALSE) or not (TRUE this is default)}

\item{options}{A list of possible plot configurations. See comments below}
}
\value{
sub A plotly widget
}
\description{
MPS data visualizer (interactive)
}
\details{
Plots intensities with corresponding allele variant for one sample. Does not yet handle replicates. Can handle RU grouping with separator grpsymbol.
}
\author{
Oyvind Bleka
}
