% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSumPH.R
\name{plotSumPH}
\alias{plotSumPH}
\title{plotSumPH}
\usage{
plotSumPH(
  samples,
  kit = NULL,
  ignoreMarkers = c("Y", "AM"),
  LUSsymbol = "_",
  MPSsymbol = ":"
)
}
\arguments{
\item{samples}{A List [[sample]][[locus]] with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{kit}{shortname of kit: Obtained from getKit()}

\item{ignoreMarkers}{Give string subset of markers that should be ignored in visualization}

\item{LUSsymbol}{Used to recognize CE in LUS alleles}

\item{MPSsymbol}{Used to recognize CE in MPS alleles}
}
\description{
Plotting sumPH per marker
}
\details{
The function is fitting gamma distribution to the sumPH per marker to investigate fit: Degradation or sample differences
}
\author{
Oyvind Bleka
}
