% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTippet.R
\name{plotTippet}
\alias{plotTippet}
\title{plotTippet}
\usage{
plotTippet(
  x,
  lr0 = NULL,
  qq = c(0.5, 0.95, 0.99),
  mtxt = "",
  xrange = NULL,
  dig = 2,
  returnStatsOnly = FALSE
)
}
\arguments{
\item{x}{log10 LR values of non-contributors}

\item{lr0}{Observed LR to superimpose (log10 LR)}

\item{qq}{Quantiles to calculate and show in figure}

\item{mtxt}{Added text to figure}

\item{xrange}{Range of x-limit (Can be provided by user)}

\item{dig}{Number of decimals to use}

\item{returnStatsOnly}{Whether to only return stats}
}
\description{
Plotting non-contributor plot (tippets)
}
\details{
This function prompts the user to obtain a filename and saves text based tables and ensures that a proper extension is used (txt or csv)
}
\author{
Oyvind Bleka
}
