% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopEPG.R
\name{plotTopEPG}
\alias{plotTopEPG}
\title{plotTopEPG}
\usage{
plotTopEPG(MLEobj, DCobj = NULL, kitname = NULL, threshT = 0)
}
\arguments{
\item{MLEobj}{An object returned from calcMLE/contLikMLE}

\item{DCobj}{An object returned from devonvolve: Must be run with same object as MLEobj}

\item{kitname}{Name of kit:  Obtained from getKit()}

\item{threshT}{The detection threshold can be shown in gray in the plot.}
}
\description{
Replaced by plotTopEPG2
}
\author{
Oyvind Bleka
}
