% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopEPG2.R
\name{plotTopEPG2}
\alias{plotTopEPG2}
\title{plotTopEPG2}
\usage{
plotTopEPG2(MLEobj, DCobj = NULL, kit = NULL, dyeYmax = TRUE,
  plotRepsOnly = TRUE, options = NULL)
}
\arguments{
\item{MLEobj}{An object returned from contLikMLE}

\item{DCobj}{An object returned from devonvolve: Must be run with same object as MLEobj}

\item{kit}{Short name of kit: See supported kits with getKit(). Argument ignored if degradation model used.}

\item{dyeYmax}{A boolean of whether the Y-axis scale should be dye specific}

\item{plotRepsOnly}{A boolean of whether only replicates are plotted}

\item{options}{Layout options can be provided: w0 (layout width),marg0 (margin), txtsize0 (text size), locsize0 (locus name size), minY (minimum Y-axis height), ymaxscale (Scaling maximum Y-axis height), transdeg (transparanct of bars 0=max, 1=min)}
}
\value{
sub A plotly widget
}
\description{
EPG data visualizer (interactive)
}
\details{
Plots the expected peak heights of the top genotypes. The peak heights for corresponding alleles (one sample) are superimposed.
}
\author{
Oyvind Bleka
}
