% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopLUS.R
\name{plotTopLUS}
\alias{plotTopLUS}
\title{plotTopLUS}
\usage{
plotTopLUS(MLEobj, DCobj = NULL, threshT = 0, LUSsymbol = "_")
}
\arguments{
\item{MLEobj}{An object returned from calcMLE/contLikMLE}

\item{DCobj}{An object returned from deconvolve: Must be run with same object as MLEobj}

\item{threshT}{The detection threshold can be shown in gray in the plot.}

\item{LUSsymbol}{The separator for each variant. First is assumed to be CE.}
}
\description{
A dataplotter for MPS data with LUS format
}
\details{
Utilizing existing function plotTopMPS2 to show.
}
\author{
Oyvind Bleka
}
