% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(
  mixData,
  refData = NULL,
  popFreq = NULL,
  minF = NULL,
  normalize = FALSE,
  threshT = NULL,
  fillHomGen = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{mixData}{A list with evidence profiles [[sample]][[locus]]$adata/hdata}

\item{refData}{A list with reference profiles [[reference]][[locus]]$adata}

\item{popFreq}{A list with population frequencies [[locus]]}

\item{minF}{The freq value included for new alleles (new alleles as potential stutters will have 0). Default NULL is using min.observed in popFreq.}

\item{normalize}{Whether normalization should be applied or not. Default is FALSE.}

\item{threshT}{A detection threshold value or thresholds per makers (marker names must be defined). NULL ignores filtering.}

\item{fillHomGen}{Whether to fill in homozygote genotypes given with one allele}

\item{verbose}{Whether printing out information}
}
\value{
Restructured data in list used as input for functions evaluating the liklihood function
}
\description{
Reorganasing data for calculation preparations
}
\details{
Optionally to use beforehand (still required for Qual. calcs)
Helpfunction to reorganising data which are input for further calculations (includes Qassignation). 
Also takes detection thresholds as argument to remove possible alleles below the thresholds.
}
\author{
Oyvind Bleka
}
