% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualLikMLE.R
\name{qualLikMLE}
\alias{qualLikMLE}
\title{qualLikMLE}
\usage{
qualLikMLE(nC, samples, popFreq, refData = NULL, condOrder = NULL,
  knownRef = NULL, prC = 0.05, fst = 0, prDcontr = NULL,
  prDcommon = NULL, steptol = 1e-06, maxIter = 100, prDv0 = c(0.1,
  0.35, 0.7))
}
\arguments{
\item{nC}{Number of contributors in model. Must be a constant.}

\item{samples}{A List with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with (2-size) allele-vector given in list element [[i]][[s]].}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model.}

\item{knownRef}{Specify known non-contributing references from refData (index). For instance knownRef=(1,2) means that reference 1 and 2 is known non-contributor in the hypothesis. This affectes coancestry correction.}

\item{prC}{A numeric for allele drop-in probability. Can be a vector (must contain the marker names). Default is 0.}

\item{fst}{The co-ancestry coefficient. Can be a vector (must contain the marker names). Default is 0.}

\item{prDcontr}{assumed known dropout parameter for all contributors, NA means to be optimized. Must be a nC long vector if given.}

\item{prDcommon}{vector indicating which contributors should share common drop-out parameter. Assign integers to contributors with common parameters. NA means not optimized.}

\item{steptol}{Argument used in the nlm function for faster return from the optimization (tradeoff is lower accuracy).}

\item{maxIter}{Maximum number of iterations for the optimization}

\item{prDv0}{Start values for fitting the drop-out probabilities (will be spanned if multidimensional)}
}
\value{
ret A list(fit,model,nDone,delta,seed,prepareC) where fit is Maximixed likelihood elements for given model.
}
\description{
Optimizing the likelihood function based on qualitative model (LRmix)
}
\details{
The R-package forensim is used to calculate the likelihood value for given data and parameters.
}
\examples{
\dontrun{
kit = "ESX17"
popfn = paste(path.package("euroformix"),"FreqDatabases",paste0(kit,"_Norway.csv")
,sep=.Platform$file.sep)
evidfn = paste(path.package("euroformix"),"examples",paste0(kit,"_3p.csv"),
sep=.Platform$file.sep)
popFreq = freqImport(popfn)[[1]] #obtain list with population frequencies
samples = sample_tableToList(tableReader(evidfn))
dat = prepareData(samples,popFreq=popFreq) #obtain data to use for analysis
fit = qualLikMLE(nC=2,samples=dat$samples,popFreq=dat$popFreq)
fit = qualLikMLE(nC=3,samples=dat$samples,popFreq=dat$popFreq,prDcommon=c(1,1,2))
}
}
\author{
Oyvind Bleka
}
