% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runexample.R
\name{runexample}
\alias{runexample}
\title{runexample}
\usage{
runexample(NOC, popfn, evidfn, reffn, POIind = 1, condInd = NULL,
  kit = NULL, modelDegrad = FALSE, modelBWstutt = FALSE,
  modelFWstutt = FALSE, findOptimalModel = TRUE, fst = 0.01,
  lambda = 0.01, prC = 0.05, threshT = 50, alpha = 0.01,
  minF = NULL, seed = 1, nDone = 2, mcmcIter = 500,
  mcmcDelta = 2, intRelTol = 0.1, intMaxEval = 1000,
  verbose = TRUE, maxThreads = 32)
}
\arguments{
\item{NOC}{Number of contributors}

\item{popfn}{Filename of population frequency file (must contain EFM compatible text file)}

\item{evidfn}{Filename of evidence(s) profile file (must contain EFM compatible text file)}

\item{reffn}{Filename of reference(s) profile file (must contain EFM compatible text file)}

\item{POIind}{Index of the references being Person of interest (POI), only Hp}

\item{condInd}{Index of conditional references (both Hp and Hd)}

\item{kit}{Kitname (shortname of kit obtained by getKit()). Used to model degradation and visualize EPG.}

\item{modelDegrad}{Whether degradation should be modeled (requires kit to be specified)}

\item{modelBWstutt}{Whether backward stutter should be modeled}

\item{modelFWstutt}{Whether forward stutter should be modeled}

\item{findOptimalModel}{Whether to find the optimal model (using AIC criterion)}

\item{fst}{The co-ancestry coefficient (theta-correction). Can be a vector (must be number of marker long)}

\item{lambda}{Parameter in modeled peak height shifted exponential model. Can be a vector (must be number of marker long)}

\item{prC}{Allele drop-in probability. Can be a vector (must be number of marker long)}

\item{threshT}{The analytical/detection threshold given. Can be a vector (must be number of marker long, and named with marker names to plot)}

\item{alpha}{The significance level used for the envelope test (validMLEmodel). Default is 0.01}

\item{minF}{Minimum frequency used in the analysis (can be specified, otherwise minimum observed is used by default)}

\item{seed}{Seed used in contLikMLE/contLikMCMC)}

\item{nDone}{Number of required optimizations in contLikMLE}

\item{mcmcIter}{Number of iterations used in contLikMCMC}

\item{mcmcDelta}{Scaling of variance used for proposal in contLikMCMC (calibrated based on mcmc$accrat object)}

\item{intRelTol}{relative tolerance for integration  (contLikINT)}

\item{intMaxEval}{max number of evalutions in integration (contLikINT)}

\item{verbose}{Boolean whether printing data (EPG) and progress}

\item{maxThreads}{Maximum number of threads to be executed by the parallelization}
}
\value{
A list with calculated result elements
}
\description{
Calculate LR based on the inbuilt methods in euroformix for a specific setting
}
\details{
Using following inbuilt methods:  sample_tableToList,prepareData,contLikMLE,logLiki,validMLEmodel,deconvolve,contLikMCMC,contLikINT, calcGjoint,plotEPG2
}
\examples{
\dontrun{
kit = "ESX17"
popfn = paste(path.package("euroformix"),"FreqDatabases",
 paste0(kit,"_Norway.csv"),sep=.Platform$file.sep)
evidfn = paste(path.package("euroformix"),"examples",paste0(kit,"_3p.csv"),
 sep=.Platform$file.sep)
reffn = paste(path.package("euroformix"),"examples",paste0(kit,"_refs.csv"),
 sep=.Platform$file.sep)
resList = runexample(NOC=3,popfn,evidfn,reffn,POIind=1,condInd=2,kit=kit,
 modelDegrad=TRUE,modelBWstutt=TRUE,verbose=TRUE)
}
}
\author{
Oyvind Bleka
}
