% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_listToTable.R
\name{sample_listToTable}
\alias{sample_listToTable}
\title{sample_listToTable}
\usage{
sample_listToTable(outL)
}
\arguments{
\item{outL}{A list on form outL[[samplename]][[locusname]]$adata,outL[[samplename]][[locusname]]$hdata (Evid or refs)}
}
\value{
table
}
\description{
Converting profiles from list to table format (helpfunction)
}
\author{
Oyvind Bleka
}
