% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_tableToList.R
\name{sample_tableToList}
\alias{sample_tableToList}
\title{sample_tableToList}
\usage{
sample_tableToList(table)
}
\arguments{
\item{table}{A table with data (Evid or refs)}
}
\value{
outL A data list
}
\description{
Converting table to list format (helpfunction)
}
\author{
Oyvind Bleka
}
