% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVecRightOrder.R
\name{setVecRightOrder}
\alias{setVecRightOrder}
\title{setVecRightOrder}
\usage{
setVecRightOrder(vec, orderedNames)
}
\arguments{
\item{vec}{A vector with values to sort by letting names(vec) have same order as orderedNames}

\item{orderedNames}{A vector with names to sort vec vector with}
}
\value{
Vector with same order as specified in orderedNames
}
\description{
Sorting vector with respect to names of an input (case insensitive)
}
\details{
Helpfunction to set vector to same locus order as output from prepareC (order based on evidence profiles)
Used in functions contLikMLE, contLikINT, calcRMPfst
}
\author{
Oyvind Bleka
}
