% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSaver.R
\name{tableSaver}
\alias{tableSaver}
\title{tableSaver}
\usage{
tableSaver(tab, sep = "txt")
}
\arguments{
\item{tab}{Table to be saved}

\item{sep}{Separator to use for saving table (decided by the program)}
}
\description{
A robust function for saving tables (interacts with gfile)
}
\details{
This function prompts the user to obtain a filename and saves text based tables and ensures that a proper extension is used (txt or csv)
}
\author{
Oyvind Bleka
}
