% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMCMC.R
\name{validMCMC}
\alias{validMCMC}
\title{validMCMC}
\usage{
validMCMC(mcmcfit, trace = TRUE, acf = TRUE, addLastMx = TRUE)
}
\arguments{
\item{mcmcfit}{A object returned by contLikMCMC}

\item{trace}{Whether showing trace of samples.}

\item{acf}{Whether showing autocorrelation function of samples.}

\item{addLastMx}{Whether showing the mixture proportion of last contributor}
}
\description{
Validates aposteriori samples from MCMC method
}
\details{
This function takes samples from the MCMC as given in a matrix and shows the aposterior functions.
}
\examples{
\dontrun{
kit = "ESX17"
sep0 = .Platform$file.sep
AT0 = 50
popfn = paste(path.package("euroformix"),"FreqDatabases",paste0(kit,"_Norway.csv"),sep=sep0)
evidfn = paste(path.package("euroformix"),"examples",paste0(kit,"_3p.csv"),sep=sep0)
reffn = paste(path.package("euroformix"),"examples",paste0(kit,"_refs.csv"),sep=sep0)
popFreq = freqImport(popfn)[[1]] #obtain list with population frequencies
samples = sample_tableToList(tableReader(evidfn))
refData = sample_tableToList(tableReader(reffn))
dat = prepareData(samples,refData=refData,popFreq=popFreq,threshT=AT) 
plotEPG2(dat$samples,dat$refData,kit=kit,AT=AT0)
mlefit = contLikMLE(3,dat$samples,dat$popFreq,dat$refData,1:3,
	kit=kit,xi=NULL,prC=0.05,lambda=0.01,seed=1)
mcmcfit = contLikMCMC(mlefit,niter=5000,delta=1,seed=1) #mcmcfit$acc
validMCMC(mcmcfit,acf=FALSE)
}
}
\author{
Oyvind Bleka
}
