% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMLEmodel.R
\name{validMLEmodel}
\alias{validMLEmodel}
\title{validMLEmodel}
\usage{
validMLEmodel(mlefit, kit = NULL, plottitle = "PP-plot",
  alpha = 0.01, createplot = TRUE, verbose = TRUE, maxThreads = 32)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE}

\item{kit}{Shortname of kit used}

\item{plottitle}{Maintitle text used in the PP-plot}

\item{alpha}{The significance level used for the envelope test. Default is 0.01}

\item{createplot}{Whether plot should be created}

\item{verbose}{Whether printing out information about significant alleles}

\item{maxThreads}{Maximum number of threads to be executed by the parallelization}
}
\value{
retinfo A dataframe with information from the calculated validation model
}
\description{
validMLEmodel makes model validation of whether the observed peak heights fits the fitted model
}
\details{
Obtaining the cumulative probability of the observed allele peaks are calculated and compared with a uniform distribution (theoretical)
Function calls a 'cumulative procedure' in C++ which utilizes paralellisation (OpenMP).
}
\examples{
\dontrun{
kit = "ESX17"
sep0 = .Platform$file.sep
AT0 = 50
popfn = paste(path.package("euroformix"),"FreqDatabases",paste0(kit,"_Norway.csv"),sep=sep0)
evidfn = paste(path.package("euroformix"),"examples",paste0(kit,"_3p.csv"),sep=sep0)
reffn = paste(path.package("euroformix"),"examples",paste0(kit,"_refs.csv"),sep=sep0)
popFreq = freqImport(popfn)[[1]] #obtain list with population frequencies
samples = sample_tableToList(tableReader(evidfn))
refData = sample_tableToList(tableReader(reffn))
dat = prepareData(samples,refData=refData,popFreq=popFreq,threshT=AT) 
plotEPG2(dat$samples,dat$refData,kit=kit,AT=AT0)
mlefit = contLikMLE(3,dat$samples,dat$popFreq,dat$refData,1:3,kit=kit,
	xi=NULL,prC=0.05,lambda=0.01,seed=1)
validMLEmodel(mlefit,kit=kit)
}
}
\author{
Oyvind Bleka
}
