% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcINT.R
\name{calcINT}
\alias{calcINT}
\title{calcINT}
\usage{
calcINT(
  mlefit,
  lower,
  upper,
  reltol = 1,
  scale = 0,
  maxEval = 1000,
  verbose = FALSE
)
}
\arguments{
\item{mlefit}{Fitted object using calcMLE}

\item{lower}{Lower bounds of parameters. Must be in following order: mx1,..,mx_(nC-1),mu,sigma,beta,xi.}

\item{upper}{Upper bounds of parameters. Must be in following order: mx1,..,mx_(nC-1),mu,sigma,beta,xi.}

\item{reltol}{Required relative tolerance error of evaluations in integration routine.}

\item{scale}{used to avoid underflow (should be the maximum likelihood value)}

\item{maxEval}{Maximum number of evaluations in the adaptIntegrate function.}

\item{verbose}{Whether printing limits to integrate over. Printing progress if maxEval>0}
}
\value{
ret A list(loglik,logdeviation,nEvals,scale) where loglik is Marginalized likelihood, logdeviation is the error-interval of loglik, nEvals is the total number of likelihood evaluations.
}
\description{
Marginalizing the likelihood through numerical integration.
}
\references{
Hahn,T. (2005). CUBA - a library for multidimensional numerical integration. Computer Physics Communications, 168(2),78-95.
}
\author{
Oyvind Bleka
}
\keyword{Marginalized}
\keyword{likelihood}
