% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRint.R
\name{calcLRint}
\alias{calcLRint}
\title{calcLRint}
\usage{
calcLRint(
  mlefitHp,
  mlefitHd,
  reltol = 1,
  maxEval = 1000,
  dev = 2,
  verbose = TRUE
)
}
\arguments{
\item{mlefitHp}{A fitted object returned from calcMLE (under Hp)}

\item{mlefitHd}{A fitted object returned from calcMLE (under Hd)}

\item{reltol}{Required relative tolerance error of evaluations in integration routine}

\item{maxEval}{Maximum number of evaluations in the integration routine function.}

\item{dev}{Scaling relative to 2*SE (i.e. 97.5 percentile). Should cover most of posterior distribution}

\item{verbose}{Whether printing simulation progress.}
}
\value{
returned inferred LR with error range (reflects relative error)
}
\description{
Calculating marginal likelihood based on integration
}
\details{
The procedure is a wrapper for the numerical integration function for obtaining Bayes Factor
}
\author{
Oyvind Bleka
}
