% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRmle.R
\name{calcLRmle}
\alias{calcLRmle}
\title{calcLRmle}
\usage{
calcLRmle(mlefitHp, mlefitHd)
}
\arguments{
\item{mlefitHp}{A fitted object returned from contLikMLE (under Hp)}

\item{mlefitHd}{A fitted object returned from contLikMLE (under Hd)}
}
\description{
A function for calculating LR from MCMC simulations
}
\details{
Returns a conservative (quantile) LR or a Full Bayesian based LR
}
