% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTippet.R
\name{calcTippet}
\alias{calcTippet}
\title{calcTippet}
\usage{
calcTippet(
  tipIdx,
  mlefitHp,
  mlefitHd,
  niter = 100,
  type = "MLE",
  LRobs = NULL,
  MLEopt = NULL,
  INTopt = list(reltol = 1, maxEval = 1000, dev = 2),
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{tipIdx}{Index in condOrder which are replaced with a random man}

\item{mlefitHp}{A fitted object returned from calcMLE (under Hp)}

\item{mlefitHd}{A fitted object returned from calcMLE (under Hd)}

\item{niter}{Number of drawn non-contributors/iterations}

\item{type}{type of Tippet analysis to conduct (MLE or INT)}

\item{LRobs}{The user can send an observed LR to superimpose to the plot}

\item{MLEopt}{MLE options (Not used since it uses what is stored in mlefit object)}

\item{INTopt}{INT options (must include reltol, maxEval, dev)}

\item{seed}{Seed used for reproducibility}

\item{verbose}{Whether printing the progress. Default is TRUE}
}
\value{
returned non-contributor LR values and statistics
}
\description{
Performing tippet/non-contributor analysis for a given set of hypotheses
}
\author{
Oyvind Bleka
}
