% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combineRank.R
\name{combineRank}
\alias{combineRank}
\title{combineRank}
\usage{
combineRank(dlist, loghdval = Inf, alpha = 0.99, maxsearch = 1000)
}
\arguments{
\item{dlist}{A list with vector elements of decreased sorted values.}

\item{loghdval}{An optional value used to stop the search together with alpha.}

\item{alpha}{An optional value used to stop the search together with loghdval.}

\item{maxsearch}{The size of the ranked deconvolved profile table will not exceed this number (used to avoid endless search).}
}
\value{
ret A list(rankG,pG) where rankG is the ranked genotypes with corresponding probabilities in pG.
}
\description{
A search algorithm which creates a ranked table of the greatest combined sum from each vector-elements in a list.
}
\details{
The procedure takes a list with vector elements of decreased sorted values and creates a ranked table of the most probable summed values (sum) from each vector-elements in the list.

The search is stopped when SUM(exp(sum-loghdval))>alpha or when the size of the ranked table is equal maxsearch
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}
\keyword{optimalisation}

