% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contLikMCMC.R
\name{contLikMCMC}
\alias{contLikMCMC}
\title{contLikMCMC}
\usage{
contLikMCMC(
  mlefit,
  niter = 10000,
  delta = 2,
  verbose = TRUE,
  seed = NULL,
  usePhi = FALSE,
  mcmcObj = NULL
)
}
\arguments{
\item{mlefit}{Fitted object using calcMLE}

\item{niter}{Number of samples in the MCMC-sampling.}

\item{delta}{A numerical parameter to scale with the covariance function Sigma. Default is 2. Should be higher to obtain lower acception rate.}

\item{verbose}{Whether printing simulation progress. Default is TRUE}

\item{seed}{The user can set seed if wanted}

\item{usePhi}{Whether the transformed domain of the parameters (phi) should be used instead. This affects the priors!}

\item{mcmcObj}{An object from contLikMCMC output}
}
\value{
ret A list (logmargL,posttheta,postlogL,logpX,accrat) where logmargL is Marginalized likelihood, posttheta is the posterior samples, postlogL is log-likelihood values
}
\description{
calcMCMC provides samples from the posterior distribution of the model parameters
}
\details{
The procedure also uses the samples to approximate the marginal likelihood used for Bayes Factor 

The Metropolis Hastings routine uses following proposal: Multivariate Normal distribution with mean 0 and covariance as delta multiplied with the inverse negative hessian with MLE inserted.
Marginalized likelihood (Bayesian) is estimated using Metropolis Hastings with the "GD-method, Gelfand and Dey (1994).
}
\author{
Oyvind Bleka
}
