% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contLikMCMCpara.R
\name{contLikMCMCpara}
\alias{contLikMCMCpara}
\title{contLikMCMCpara}
\usage{
contLikMCMCpara(mlefit, niter = 10000, delta = 10, maxxi = 1)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE}

\item{niter}{Number of samples in the MCMC-sampling.}

\item{delta}{A numerical parameter to scale with the covariance function Sigma. Default is 2. Should be higher to obtain lower acception rate.}

\item{maxxi}{Upper boundary of the xi-parameters}
}
\value{
ret A list (margL,posttheta,postlogL,logpX,accrat,Ubound ) where margL is Marginalized likelihood for hypothesis (model) given observed evidence, posttheta is the posterior samples from a MC routine, postlogL is sampled log-likelihood values, accrat is ratio of accepted samples. Ubound is upper boundary of parameters.
}
\description{
Parallelization on contLikMCMC using snow
}
\details{
The procedure is doing parallelization of the contLikMCMC function
}
\references{
Craiu,R.V. and Rosenthal, J.S. (2014). Bayesian Computation Via Markov Chain Monte Carlo. Annu. Rev. Stat. Appl., 1,179-201.
}
\author{
Oyvind Bleka
}
\keyword{BayesianModels,}
\keyword{MCMC,}
\keyword{MarginalizedLikelihoodEstimation}
\keyword{MetropolisHastings,}
\keyword{continuous,}
