% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contLikSearch.R
\name{contLikSearch}
\alias{contLikSearch}
\title{contLikSearch}
\usage{
contLikSearch(
  NOC = 2:3,
  modelDegrad = FALSE,
  modelBWstutt = FALSE,
  modelFWstutt = FALSE,
  samples = NULL,
  popFreq = NULL,
  refData = NULL,
  condOrder = NULL,
  knownRefPOI = NULL,
  prC = 0,
  nDone = 3,
  threshT = 50,
  fst = 0,
  lambda = 0,
  delta = 1,
  kit = NULL,
  verbose = TRUE,
  difftol = 0.01,
  knownRel = NULL,
  ibd = c(1, 0, 0),
  minF = NULL,
  normalize = TRUE,
  priorBWS = NULL,
  priorFWS = NULL,
  seed = NULL,
  maxThreads = 0,
  alpha = 0.01,
  steptol = 1e-04,
  adjQbp = FALSE,
  resttol = 1e-06
)
}
\arguments{
\item{NOC}{#The vector of number of contributors to search}

\item{modelDegrad}{Whether degradation should be modeled (requires kit to be specified): can be vector (FALSE,TRUE)}

\item{modelBWstutt}{Whether backward stutter should be modeled: can be vector (FALSE,TRUE)}

\item{modelFWstutt}{Whether forward stutter should be modeled: can be vector (FALSE,TRUE)}

\item{samples}{A List with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with (2-size) allele-vector given in list element [[i]][[s]].}

\item{condOrder}{condOrder under Hd. See calcMLE function for more details.}

\item{knownRefPOI}{Specify index of known POI under Hp (can only be one number)}

\item{prC}{A numeric for allele drop-in probability. Default is 0.}

\item{nDone}{Number of optimizations required providing equivalent results (same logLik value obtained)}

\item{threshT}{The detection threshold given. Used when considering probability of allele drop-outs.}

\item{fst}{The co-ancestry coefficient. Default is 0.}

\item{lambda}{Parameter in modeled peak height shifted exponential model. Default is 0.}

\item{delta}{Scaling of variation of normal distribution when drawing random startpoints. Default is 1.}

\item{kit}{Used to model degradation. Must be one of the shortnames of kit: check getKit()}

\item{verbose}{Whether printing optimization progress. Default is TRUE.}

\item{difftol}{Tolerance for being exact in log-likelihood value (relevant when nDone>1)}

\item{knownRel}{gives the index of the reference which the 1st unknown is related to.}

\item{ibd}{the identical by decent coefficients of the relationship (specifies the type of relationship)}

\item{minF}{The freq value included for new alleles (new alleles as potential stutters will have 0). Default NULL is using min.observed in popFreq.}

\item{normalize}{Whether normalization should be applied or not. Default is FALSE.}

\item{priorBWS}{Prior function for BWS-parameter. Flat prior on [0,1] is default.}

\item{priorFWS}{Prior function for FWS-parameter. Flat prior on [0,1] is default.}

\item{seed}{The user can set seed if wanted}

\item{maxThreads}{Maximum number of threads to be executed by the parallelization}

\item{alpha}{The significance level used for the envelope test (validMLEmodel). Default is 0.01}

\item{steptol}{Argument used in the nlm function for faster return from the optimization (tradeoff is lower accuracy).}

\item{adjQbp}{Indicate whether fragmenth length of Q-allele is based on averaged weighted with frequencies}

\item{resttol}{Restriction tolerance used to restrict genotype outcome. 0=No restriction, 1=Full restriction.}
}
\value{
A list with fitted models, and table results for the defined model outcome
}
\description{
Searches through a set of model configurations which optimizes the likelihood function
}
\author{
Oyvind Bleka
}
