% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{deconvolve}
\usage{
deconvolve(mlefit, alpha = 0.95, maxlist = 1000, verbose = TRUE)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE function.}

\item{alpha}{Required sum of the listed posterior probabilities.}

\item{maxlist}{The ranked deconvolved profile list will not exceed this number (used to avoid endless search).}

\item{verbose}{Whether printing deconvolution progress. Default is TRUE.}
}
\value{
ret A list(table1,table2,table3,table4,rankGi,rankG,pG) where rankG is the ranked genotypes with corresponding probabilities in pG. rankgGi is the same, but per marker. table1 is rankG and pG combined (joint results). table2 uses rankGi to find marginal results for top-genotypes. table3 and table4 shows this marginalized on genotypes and alleles per contributor
}
\description{
deconvolve calculates the posterior probability of each genotypes based on a fitted MLE model
}
\details{
The procedure calculates the likelihood for each single locus.
}
\references{
Cowell,R.G. et.al. (2014). Analysis of forensic DNA mixtures with artefacts. Applied Statistics, 64(1),1-32.
}
\author{
Oyvind Bleka
}
\keyword{deconvolution}
