% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/efm.R
\name{efm}
\alias{efm}
\title{efm}
\usage{
efm(envirfile = NULL)
}
\arguments{
\item{envirfile}{A file to a saved environment of a project}
}
\description{
efm (EuroForMix) is a GUI wrapper for euroformix
}
\details{
The function is a graphical layer for the functions in the package euroformix. See ?euroformix for more information.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

