% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactMACdistr.R
\name{exactMACdistr}
\alias{exactMACdistr}
\title{exactMACdistr}
\usage{
exactMACdistr(si)
}
\arguments{
\item{si}{A vector with each element as the sum of allele frequencies for observed alleles.}
}
\value{
ret A vector with false positive probabilities for each given number of observed alleles.
}
\description{
Calculation of false positive probability of Matching allele counter.
}
\details{
This function will calculate the false positive probability for a given range of allele matches.
Thanks 1: Thore Egeland: Explicit formulas.
Thanks 2: Myself to have updated the function
}
\author{
Oyvind Bleka
}
