% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgammamodel.R
\name{fitgammamodel}
\alias{fitgammamodel}
\title{fitgammamodel}
\usage{
fitgammamodel(y, x = NULL, DEG = TRUE, niter = 30, delta = 1,
  plott = FALSE, alpha = 0.05, offset = 125, scale = 100)
}
\arguments{
\item{y}{a vector of sum peak heights per loci}

\item{x}{a vector of base pairs}

\item{DEG}{a boolean of whether degradation (i.e. the bp at x-axis should be considered)}

\item{niter}{number of random samples}

\item{delta}{Standard deviation of normal distribution when drawing random startpoints. Default is 1.}

\item{plott}{a boolean of whether plotting fitted model along with data}

\item{alpha}{The alpha/2 and 1-alpha/2 percentiles will be plotted if x is provided}

\item{offset}{a number giving shift of bp in degradation model.}

\item{scale}{a number giving scaling of shifted bp in degradation model.}
}
\value{
Estimated parameters (mean,coefVar,slopePara)
}
\description{
fitgammamodel is a function to fit sumPH of the observed peak heights to a gamma model
}
\details{
This function is used for fitting a gamma regression model
}
\examples{
th = c(1000,0.8,0.6)
n = 100 #number of samples
x = seq(10,300,l=n)
y = rgamma(n,shape=2/th[2]^2*th[3]^((x-125)/100),scale=th[1]*th[2]^2)
fitgammamodel(y,x,DEG=TRUE,niter=10,plott=TRUE,alpha=0.05)
fitgammamodel(y,x,DEG=FALSE,niter=10,plott=TRUE,alpha=0.05)
}
\author{
Oyvind Bleka
}
