% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqImport.R
\name{freqImport}
\alias{freqImport}
\title{freqImport}
\usage{
freqImport(f = NULL, url = FALSE, xml = FALSE)
}
\arguments{
\item{f}{The path of file(s) or an url. If file(s), then the url argument must be FALSE.}

\item{url}{A boolean of whether f is pointing to an url or not. Default is FALSE.}

\item{xml}{A boolean of whether f is pointing to an xml format file or not. If FALSE then table formats with standard separator types are considered. Default is FALSE.}
}
\value{
dblist A list giving the population frequencies for every population
}
\description{
FreqImport is a function to import allele frequencies directly from file(s) supporting text format, xml. It also supports an URL path reading XML format (as for the STRidER webpage).
}
\details{
The function reads XML as input, directly from an URL or local file, and return a list of populations.
}
\author{
Oyvind Bleka
}
