% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genDataset.R
\name{genDataset}
\alias{genDataset}
\title{genDataset}
\usage{
genDataset(nC, popFreq, mu = 1000, sigma = 0.1, sorted = FALSE,
  threshT = 50, refData = NULL, mx = NULL, nrep = 1, stutt = 0,
  prC = 0, lambda = 0, beta = 1, kit = NULL)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{mu}{Expected peak heights for a het. single contributor allele}

\item{sigma}{Coeffecient of variance of peak heights.}

\item{sorted}{Boolean for wheter sorting the contributors with respect to decreasingly mixture proportions.}

\item{threshT}{Required allele peak height in mixture}

\item{refData}{A list of given reference profiles given as refData[[i]][[s]]. Default is random from population.}

\item{mx}{A vector of known mixture proportions. Default is random uniform.}

\item{nrep}{Number of peak height replicates (same contributors) to generate. Default is 1.}

\item{stutt}{A numerical stutter ratio (one ahead). Default is 0.}

\item{prC}{A numerical dropin probability. Default is 0.}

\item{lambda}{The rate parameter in the exponential distribution for simulating drop-in peak heights. Default is 0.}

\item{beta}{The degradation slope parameter used for simulating degradation trend (requires valid kit to be specified). Default is 1.}

\item{kit}{shortname of kit: {"ESX17","ESI17","ESI17Fast","ESX17Fast","Y23","Identifiler","NGM","ESSPlex","ESSplexSE","NGMSElect","SGMPlus","ESX16", "Fusion","GlobalFiler"}}
}
\value{
List with elements theta,samples,refData where theta is the true parameters of the model. samples is a list with samples which for each samples has locus-list elements with list elements adata and hdata
}
\description{
Function for generating replicated mixture samples given same contributors and model parameters.
}
\details{
genDataset samples random mixture peak heights given as gamma(rho*sum(h_k),tau), with h_k as peak height of k-te contributor.
genData conditions on alleles given by refData. Empty references are generated with population frequencies.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

