% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGlist.R
\name{getGlist}
\alias{getGlist}
\title{getGlist}
\usage{
getGlist(popFreq)
}
\arguments{
\item{popFreq}{A list of allele frequencies for a given population. Each named element in the list must be a allele-named vector with allele frequencies.}
}
\value{
Glist A list with genotypes and genotype probabilities for each locus.
}
\description{
getGlist Returns a list of genotypes with corresponding probabilities for a given allele frequency-list.
}
\details{
The function returns the list of all possible observed genotypes, with corresponding Hardy-Weinberg assumed probabilities. The allele-names in popFreq needs to be numbers.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

