% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKit.R
\name{getKit}
\alias{getKit}
\title{getKit}
\usage{
getKit(kit = NULL, what = NA)
}
\arguments{
\item{kit}{shortname of kit: {"ESX17","ESI17","ESI17Fast","ESX17Fast","Y23","Identifiler","NGM","ESSPlex","ESSplexSE","NGMSElect","SGMPlus","ESX16", "Fusion","GlobalFiler"}}

\item{what}{options: {"Index","Panel","Short.Name","Full.Name","Marker","Allele","Size","Virtual","Color","Repeat","Range","Offset","Gender"}}
}
\value{
res A data frame with kit information
}
\description{
Function to get kit information. Tided up by Øyvind Bleka
}
\details{
Returns kit information
}
\author{
Oskar Hansson
}

