% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleType.R
\name{getSampleType}
\alias{getSampleType}
\title{getSampleType}
\usage{
getSampleType(samples, kit = NULL, LUSsymbol = "_")
}
\arguments{
\item{samples}{A list of samples (evidence) with structure [[samplename]][[locus]] = list(adata,hdata)}

\item{kit}{Kitname returned from getKit()}

\item{LUSsymbol}{Symbol to be recognized as a LUS allele}
}
\description{
A function for creating report based on CaseSolver object
}
\details{
The function returns the names of the selected profiles (when window is closed)
}
