% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iszerolik.R
\name{iszerolik}
\alias{iszerolik}
\title{iszerolik}
\usage{
iszerolik(evid, ref, nU, xi = 0)
}
\arguments{
\item{evid}{A vector with allele names}

\item{ref}{A vector with conditioned alleles in given hypothesis}

\item{nU}{Number of unknown individuals in given hypothesis}
}
\value{
TRUE/FALSE A boolean whether the likelihood vil be zero.
}
\description{
iszerolik Determine if the likelihood becomes zero
}
\details{
The function is determining whether likelihood vil be 0 (TRUE or FALSE) assuming zero drop-in
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

