% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLiki.R
\name{logLiki}
\alias{logLiki}
\title{logLiki}
\usage{
logLiki(mlefit, verbose = FALSE, maxThreads = 32)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE}

\item{verbose}{Boolean whether printing deconvolution progress. Default is TRUE.}

\item{maxThreads}{Maximum number of threads to be executed by the parallelization}
}
\value{
ret A vector with log-likelihood-values for each locus for given model
}
\description{
logLiki returns the likelihood for each markers (based on the MLE results)
}
\author{
Oyvind Bleka
}
