% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLiki.R
\name{logLiki}
\alias{logLiki}
\title{logLiki}
\usage{
logLiki(mlefit)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE}
}
\value{
ret A vector with log-likelihood-values for each locus for given model
}
\description{
logLiki calculates the likelihood of each marker of the STR DNA mixture given a theta
}
\details{
Function calls procedure in c++ by using the package Armadillo and Boost.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

