% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/noncontrMLE.R
\name{noncontrMLE}
\alias{noncontrMLE}
\title{noncontrMLE}
\usage{
noncontrMLE(mlefit, tipind = NULL, ntippet = 100)
}
\arguments{
\item{mlefit}{A fitted mle object from contLikMLE function}

\item{tipind}{Index in condOrder which are replaced with a random man}

\item{ntippet}{Number of samples to draw}
}
\value{
TRUE/FALSE A boolean whether the likelihood vil be zero.
}
\description{
noncontrMLE Calculate random log-likelhood values by exchanging a reference with a random man from population
}
\details{
The function simulates ntippet numbder of
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

