% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLUS.R
\name{plotLUS}
\alias{plotLUS}
\title{plotLUS}
\usage{
plotLUS(mixData, sn = "", refData = NULL, threshT = 0,
  LUSsymbol = "_")
}
\arguments{
\item{mixData}{List of adata- and hdata-elements. adata must be separated with LUSsymbol}

\item{sn}{The Sample name, will be plotted in title.}

\item{refData}{condition on a list$refname$locname$adata of reference alleles which are labeled in EPG}

\item{threshT}{The detection threshold can be shown in gray in the plot.}

\item{LUSsymbol}{The separator for each allele. Assumed to be "_" separating sequence variant and LUS.}
}
\description{
A dataplotter for MPS data
}
\details{
Plots peak height with corresponding allele variant for one sample
}
\author{
Oyvind Bleka
}
