% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSumPH.R
\name{plotSumPH}
\alias{plotSumPH}
\title{plotSumPH}
\usage{
plotSumPH(
  samples,
  kit = NULL,
  ignoreMarkers = c("Y", "AM"),
  LUSsymbol = "_",
  MPSsymbol = ":"
)
}
\arguments{
\item{samples}{List object of [[sampleName]][[locus]] =list(adata,hdata), Allele vector adata (can be strings or numeric), intensity vector hdata (must be numeric)}

\item{kit}{shortname of kit: Obtained from getKit()}

\item{ignoreMarkers}{Give string subset of markers that should be ignored in visualization}

\item{LUSsymbol}{Used to recognize CE in LUS alleles}

\item{MPSsymbol}{Used to recognize CE in MPS alleles}
}
\description{
Plotting sumPH per marker
}
\details{
The function is fitting gamma distribution to the sumPH per marker to investigate fit: Degradation or sample differences
}
\author{
Oyvind Bleka
}
