% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopEPG.R
\name{plotTopEPG}
\alias{plotTopEPG}
\title{plotTopEPG}
\usage{
plotTopEPG(MLEobj, DCobj = NULL, kitname = NULL, threshT = 0)
}
\arguments{
\item{MLEobj}{An object returned from contLikMLE}

\item{DCobj}{An object returned from devonvolve: Must be run with same object as MLEobj}

\item{kitname}{Name of kit: {"ESX17","ESI17","ESI17Fast","ESX17Fast","Y23","Identifiler","NGM","ESSPlex","ESSplexSE","NGMSElect","SGMPlus","ESX16", "Fusion","GlobalFiler"}}

\item{threshT}{The detection threshold can be shown in gray in the plot.}
}
\description{
EPG plotter created by Oskar Hansson.
}
\details{
Plots the expected peak heights of the top genotypes. The peak heights for corresponding alleles (one sample) are superimposed.
}
\author{
Oyvind Bleka
}
