% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopLUS.R
\name{plotTopLUS}
\alias{plotTopLUS}
\title{plotTopLUS}
\usage{
plotTopLUS(MLEobj, DCobj = NULL, threshT = 0, LUSsymbol = "_")
}
\arguments{
\item{MLEobj}{An object returned from contLikMLE}

\item{DCobj}{An object returned from devonvolve: Must be run with same object as MLEobj}

\item{threshT}{The detection threshold can be shown in gray in the plot.}

\item{LUSsymbol}{The separator for each allele. Assumed to be "_" separating sequence variant and LUS.}
}
\description{
A dataplotter for MPS data with LUS format
}
\details{
Plots the expected coverage reads of the top genotypes. The coverage reads for corresponding alleles (one sample) are superimposed.
}
\author{
Oyvind Bleka
}
