% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareC.R
\name{prepareC}
\alias{prepareC}
\title{prepareC}
\usage{
prepareC(nC, samples, popFreq, refData = NULL, condOrder = NULL,
  knownRef = NULL, kit = NULL)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{samples}{A List with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]][[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model.}

\item{knownRef}{Specify known non-contributing references from refData (index). For instance knownRef=(1,2) means that reference 1 and 2 is known non-contributor in the hypothesis. This affectes coancestry correction.}

\item{kit}{shortname of kit: {"ESX17","ESI17","ESI17Fast","ESX17Fast","Y23","Identifiler","NGM","ESSPlex","ESSplexSE","NGMSElect","SGMPlus","ESX16", "Fusion","GlobalFiler"}}
}
\value{
ret A list of data input to call the C-code with
}
\description{
prepareC is used in the functions contLikMLE,contLikMarg and contLikMCMC to prepare input to C-call
}
\details{
The function builds the data input to the C-code
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

