% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qassignate.R
\name{Qassignate}
\alias{Qassignate}
\title{Qassignate}
\usage{
Qassignate(samples, popFreq, refData = NULL, doQ = TRUE, incR = TRUE,
  incS = FALSE, minF = NULL, normalize = FALSE)
}
\arguments{
\item{samples}{A List with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{doQ}{A boolean whether to Q-designate or not}

\item{incR}{A boolean whether to include reference-alleles in the Q-assignation}

\item{incS}{A boolean whether to include potential stutters in the Q-assignation}

\item{minF}{The freq value included for new alleles (new alleles as potential stutters will have 0). Default NULL is using min.observed in popFreq.}

\item{normalize}{A boolean of whether normalization should be applied or not. Default is TRUE.}
}
\value{
ret A list(popFreq,refData,samples) with Q-designated alleles.
}
\description{
Q-designation. It also takes care of include new alleles into popFreq by assigning it lowest observed frequnce.
}
\details{
Assignes non-shown alleles as one single allele.
}
\author{
Oyvind Bleka
}
