% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Qassignate.R
\name{Qassignate}
\alias{Qassignate}
\title{Qassignate}
\usage{
Qassignate(samples, popFreq, refData = NULL, doQ = TRUE, incS = TRUE,
  incR = TRUE)
}
\arguments{
\item{samples}{A List with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{doQ}{A boolean whether to Q-designate or not}

\item{incS}{A boolean whether to include potential stutters in the Q-assignation}

\item{incR}{A boolean whether to include reference-alleles in the Q-assignation}
}
\value{
ret A list(popFreq,refData) with Q-designated alleles.
}
\description{
Q-designation. It also takes care of include new alleles into popFreq by assigning it lowest observed frequnce.
}
\details{
Assignes non-shown alleles as one single allele.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

