% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableReader.R
\name{tableReader}
\alias{tableReader}
\title{tableReader}
\usage{
tableReader(filename, header = TRUE)
}
\description{
A robust function for reading tables
}
\details{
This function reads text based tables and recognizes which of the three common separates that are used.
}
\author{
Oyvind Bleka
}
