% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validMLEmodel.R
\name{validMLEmodel}
\alias{validMLEmodel}
\title{validMLEmodel}
\usage{
validMLEmodel(mlefit, kit = NULL)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE}

\item{kit}{Shortname of kit used}
}
\value{
ret A vector for each marker with cumulative probabilities
}
\description{
validMLEmodel makes model validation whether the observed peak heights fits the maximum likelihood fitted gamma distribution.
}
\details{
The cumulative probability of the observed allele peaks are calculated and compared with a uniform distribution.
Function calls procedure in c++ by using the package Armadillo and Boost.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

